hw06, annotation and reflection. Step 3, moved exceptions classes in another package. Remove exceptions from afterMethods
