+++ hw06-annotation/src/main/java/ru/otus/merets/exceptions/TestStarterDefaultConstructorException.java
@@ -1,0 +1,7 @@
+package ru.otus.merets.exceptions;
+
+public class TestStarterDefaultConstructorException extends Exception {
+    public TestStarterDefaultConstructorException(String message) {
+        super(message);
+    }
+}
