+++ hw06-annotation/src/main/java/ru/otus/merets/exceptions/TestStarterInitFailedException.java
@@ -1,0 +1,10 @@
+package ru.otus.merets.exceptions;
+
+public class TestStarterInitFailedException extends Exception {
+    public TestStarterInitFailedException(String message) {
+        super(message);
+    }
+}
+
+
+
