+++ hw06-annotation/src/main/java/ru/otus/merets/exceptions/TestStarterRunAfterException.java
@@ -1,0 +1,7 @@
+package ru.otus.merets.exceptions;
+
+public class TestStarterRunAfterException extends Exception {
+    public TestStarterRunAfterException(String message) {
+        super(message);
+    }
+}
