+++ hw06-annotation/src/main/java/ru/otus/merets/exceptions/TestStarterRunBeforeException.java
@@ -1,0 +1,7 @@
+package ru.otus.merets.exceptions;
+
+public class TestStarterRunBeforeException extends Exception {
+    public TestStarterRunBeforeException(String message) {
+        super(message);
+    }
+}
